+++ Codewars/src/main/java/com/codecool/codewars/parseintreloaded/multiplier/Multiplier.java
@@ -1,4 +1,4 @@
 package com.codecool.codewars.parseintreloaded.multiplier;
 
-public record Multiplier(String name, int number) {
+public record Multiplier(String name, long number) {
 }
